Fix verify lint failures
