+++ src/renderer/src/components/stats/OpenCodeUsagePane.tsx
@@ -27,7 +27,9 @@ import {
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '../ui/tooltip'
 import { ClaudeUsageLoadingState } from './ClaudeUsageLoadingState'
 import { CodexUsageDailyChart } from './CodexUsageDailyChart'
+import { OpenCodeUsageRecentSessionsTable } from './OpenCodeUsageRecentSessionsTable'
 import { StatCard } from './StatCard'
+import { formatCost, formatTokens, formatUpdatedAt } from './usage-formatters'
 import { translate } from '@/i18n/i18n'
 
 const RANGE_OPTIONS: OpenCodeUsageRange[] = ['7d', '30d', '90d', 'all']
@@ -49,43 +51,6 @@ const RANGE_LABELS: Record<OpenCodeUsageRange, string> = {
   '30d': 'Last 30 days',
   '90d': 'Last 90 days',
   all: 'All time'
-}
-
-function formatTokens(value: number): string {
-  if (value >= 1_000_000) {
-    return `${(value / 1_000_000).toFixed(1)}M`
-  }
-  if (value >= 1_000) {
-    return `${(value / 1_000).toFixed(1)}k`
-  }
-  return value.toLocaleString()
-}
-
-function formatCost(value: number | null): string {
-  if (value === null) {
-    return 'n/a'
-  }
-  return value < 0.01 ? `$${value.toFixed(4)}` : `$${value.toFixed(2)}`
-}
-
-function formatUpdatedAt(timestamp: number | null): string {
-  if (!timestamp) {
-    return 'Not scanned yet'
-  }
-  return `Updated ${new Date(timestamp).toLocaleString()}`
-}
-
-function formatSessionTime(timestamp: string): string {
-  const parsed = new Date(timestamp)
-  if (Number.isNaN(parsed.getTime())) {
-    return timestamp
-  }
-  return parsed.toLocaleString(undefined, {
-    month: 'short',
-    day: 'numeric',
-    hour: 'numeric',
-    minute: '2-digit'
-  })
 }
 
 export function OpenCodeUsagePane(): React.JSX.Element {
@@ -422,78 +387,7 @@ export function OpenCodeUsagePane(): React.JSX.Element {
             </section>
           </div>
 
-          <section className="rounded-lg border border-border/60 bg-card/40 p-4">
-            <div className="mb-3">
-              <h4 className="text-sm font-semibold text-foreground">
-                {translate('auto.components.stats.OpenCodeUsagePane.4799177b1c', 'Recent sessions')}
-              </h4>
-              <p className="text-xs text-muted-foreground">
-                {translate(
-                  'auto.components.stats.OpenCodeUsagePane.81817a641a',
-                  'Most recent local OpenCode sessions in this scope.'
-                )}
-              </p>
-            </div>
-            <div className="overflow-x-auto">
-              <table className="min-w-full text-sm">
-                <thead>
-                  <tr className="border-b border-border/60 text-left text-xs text-muted-foreground">
-                    <th className="px-2 py-2 font-medium">
-                      {translate(
-                        'auto.components.stats.OpenCodeUsagePane.d97bdf6e27',
-                        'Last active'
-                      )}
-                    </th>
-                    <th className="px-2 py-2 font-medium">
-                      {translate('auto.components.stats.OpenCodeUsagePane.a4738de041', 'Project')}
-                    </th>
-                    <th className="px-2 py-2 font-medium">
-                      {translate('auto.components.stats.OpenCodeUsagePane.08c78441b7', 'Model')}
-                    </th>
-                    <th className="px-2 py-2 font-medium">
-                      {translate('auto.components.stats.OpenCodeUsagePane.d416f5cf92', 'Events')}
-                    </th>
-                    <th className="px-2 py-2 font-medium">
-                      {translate('auto.components.stats.OpenCodeUsagePane.0f2f266c9d', 'Input')}
-                    </th>
-                    <th className="px-2 py-2 font-medium">
-                      {translate('auto.components.stats.OpenCodeUsagePane.dfc4513657', 'Output')}
-                    </th>
-                    <th className="px-2 py-2 font-medium">
-                      {translate('auto.components.stats.OpenCodeUsagePane.349f7c3f5c', 'Total')}
-                    </th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {recentSessions.map((row) => (
-                    <tr key={row.sessionId} className="border-b border-border/40 last:border-b-0">
-                      <td className="px-2 py-2 text-muted-foreground">
-                        {formatSessionTime(row.lastActiveAt)}
-                      </td>
-                      <td className="px-2 py-2 text-foreground">{row.projectLabel}</td>
-                      <td className="px-2 py-2 text-muted-foreground">
-                        {row.model ??
-                          translate(
-                            'auto.components.stats.OpenCodeUsagePane.362231082f',
-                            'Unknown'
-                          )}
-                      </td>
-                      <td className="px-2 py-2 text-muted-foreground">{row.events}</td>
-                      <td className="px-2 py-2 text-muted-foreground">
-                        {formatTokens(row.inputTokens)}
-                      </td>
-                      <td className="px-2 py-2 text-muted-foreground">
-                        {formatTokens(row.outputTokens)}
-                      </td>
-                      <td className="px-2 py-2 text-muted-foreground">
-                        {formatTokens(row.totalTokens)}
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </section>
+          <OpenCodeUsageRecentSessionsTable recentSessions={recentSessions} />
         </>
       )}
     </div>
